requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cloudinary==1.16.0
cryptography==2.7
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
freeze==1.0.10
jwt==0.6.1
Markdown==3.1.1
mock==3.0.5
Pillow==6.1.0
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
psycopg2==2.4.2